+++ branch/src/main/scala/dev/alteration/branch/macaroni/extensions/Extensions.scala
@@ -1,0 +1,4 @@
+package dev.alteration.branch.macaroni.extensions
+
+export GenericExtensions.*
+export PathExtensions.*
