+++ branch/src/main/scala/dev/alteration/branch/macaroni/extensions/GenericExtensions.scala
@@ -1,0 +1,11 @@
+package dev.alteration.branch.macaroni.extensions
+
+import dev.alteration.branch.lzy.Lazy
+
+object GenericExtensions {
+
+  extension [A](a: A) {
+    def lzy(): Lazy[A] = Lazy.fn(a)
+  }
+
+}
