+++ branch/src/main/scala/dev/alteration/branch/macaroni/extensions/StringContextExtensions.scala
@@ -1,0 +1,15 @@
+package dev.alteration.branch.macaroni.extensions
+
+object StringContextExtensions {
+
+  extension (sc: StringContext) {
+
+    /** Case-insensitive regex string interpolation
+      *
+      * @return
+      *   A Regex that matches the interpolated string case-insensitively
+      */
+    def ci = ("(?i)" + sc.parts.mkString).r
+  }
+
+}
